+++ src/services/pr.js
@@ -1,0 +1,150 @@
+const fs = require("fs").promises;
+const path = require("path");
+const inquirer = require("inquirer");
+const clipboardy = require("clipboardy");
+const { COMMIT_TYPES } = require("../constants");
+const { executeCommand } = require("../utils/helpers");
+
+/**
+ * Checks for Pull Request templates in the `.github` folder.
+ * @returns {Promise<string[]>} An array of template file paths.
+ */
+async function getPRTemplates() {
+  const githubPath = path.join(process.cwd(), ".github");
+  const templateDirPath = path.join(githubPath, "PULL_REQUEST_TEMPLATE");
+  const templates = [];
+
+  try {
+    const templateDirExists = await fs
+      .stat(templateDirPath)
+      .then((stat) => stat.isDirectory())
+      .catch(() => false);
+    if (templateDirExists) {
+      const files = await fs.readdir(templateDirPath);
+      for (const file of files) {
+        if (file.endsWith(".md")) {
+          templates.push(path.join(templateDirPath, file));
+        }
+      }
+    }
+
+    if (templates.length === 0) {
+      const githubDirExists = await fs
+        .stat(githubPath)
+        .then((stat) => stat.isDirectory())
+        .catch(() => false);
+      if (githubDirExists) {
+        const files = await fs.readdir(githubPath);
+        for (const file of files) {
+          if (
+            file.endsWith(".md") &&
+            file.toLowerCase().includes("pull_request_template")
+          ) {
+            templates.push(path.join(githubPath, file));
+          }
+        }
+      }
+    }
+  } catch (error) {}
+  return templates;
+}
+
+/**
+ * Prompts the user to choose a PR template.
+ * @param {string[]} templates An array of template file paths.
+ * @returns {Promise<string|null>} The content of the chosen template.
+ */
+async function chooseTemplate(templates) {
+  if (templates.length === 0) {
+    return null;
+  }
+
+  if (templates.length === 1) {
+    console.log(
+      `Automatically selecting the only available template: ${path.basename(
+        templates[0]
+      )}`
+    );
+    return fs.readFile(templates[0], "utf-8");
+  }
+
+  const choices = templates.map((tplPath) => ({
+    name: path.basename(tplPath),
+    value: tplPath,
+  }));
+
+  const { selectedTemplatePath } = await inquirer.default.prompt([
+    {
+      type: "list",
+      name: "selectedTemplatePath",
+      message: "Select a PR template:",
+      choices: [{ name: "No template", value: null }, ...choices],
+      default: null,
+    },
+  ]);
+
+  if (selectedTemplatePath) {
+    return fs.readFile(selectedTemplatePath, "utf-8");
+  }
+  return null;
+}
+
+/**
+ * Generates the PR description based on categorized commits and template.
+ * @param {Object.<string, string[]>} categorizedCommits Categorized commit messages.
+ * @param {string|null} templateContent Optional content from a PR template.
+ * @returns {string} The formatted PR description.
+ */
+function generatePRDescription(categorizedCommits, templateContent = null) {
+  let prBody = "";
+
+  if (templateContent) {
+    prBody += templateContent + "\n\n---\n\n";
+  }
+
+  for (const section in COMMIT_TYPES) {
+    const sectionTitle = COMMIT_TYPES[section];
+    if (
+      categorizedCommits[sectionTitle] &&
+      categorizedCommits[sectionTitle].length > 0
+    ) {
+      prBody += `### ${sectionTitle}\n\n`;
+      prBody += categorizedCommits[sectionTitle].join("\n") + "\n\n";
+    }
+  }
+
+  if (
+    categorizedCommits["Other Changes"] &&
+    categorizedCommits["Other Changes"].length > 0
+  ) {
+    prBody += `### Other Changes\n\n`;
+    prBody += categorizedCommits["Other Changes"].join("\n") + "\n\n";
+  }
+
+  return prBody.trim();
+}
+
+/**
+ * Fetches the current PR description from an existing PR using GitHub CLI.
+ * @param {string} branchName The branch name to check for existing PR.
+ * @returns {Promise<string|null>} The current PR body or null if no PR exists.
+ */
+async function getExistingPRDescription(branchName) {
+  try {
+    const prBody = await executeCommand(
+      `gh pr view ${branchName} --json body --jq .body`,
+      `Fetching existing PR description for branch "${branchName}"...`,
+      false
+    );
+    return prBody || null;
+  } catch (error) {
+    return null;
+  }
+}
+
+module.exports = {
+  getPRTemplates,
+  chooseTemplate,
+  generatePRDescription,
+  getExistingPRDescription,
+};
